Modularize Element Creation and Display Priority Flag
