+++ src/view/projects.js
@@ -3,6 +3,9 @@
   This module manages changes to the DOM relevant to projects
 */
 
+import createTitle from './shared/title.js'
+import createButton from './shared/button.js'
+
 const projects = (function () {
 
   function createProject(data) {
@@ -12,13 +15,17 @@ const projects = (function () {
     project.setAttribute('data-type', 'project');
     project.setAttribute('data-id', data.id);
 
-    const title = _createTitle(data.title, data.id);
-    const archiveBtn = _createButton(
+    const title = createTitle(
+      data.title,
+      'project',
+      data.id
+      );
+    const archiveBtn = createButton(
       `<span class="material-icons" data-id="${data.id}">inventory_2</span>`,
       'archive',
       data.id
       );
-    const deleteBtn = _createButton(
+    const deleteBtn = createButton(
       `<span class="material-icons" data-id="${data.id}">cancel</span>`,
       'delete',
       data.id
@@ -28,31 +35,6 @@ const projects = (function () {
 
     return project
   }
-
-
-  function _createTitle(value, id) {
-
-    const title = document.createElement('p');
-    title.classList.add('project-title');
-    title.setAttribute('data-type', 'project');
-    title.setAttribute('data-id', id);
-    title.innerText = value;
-
-    return title;
-
-  }
-
-
-  function _createButton(img, type, id) {
-    const btn = document.createElement('div');
-    btn.classList.add(type);
-    btn.classList.add('btn');
-    btn.setAttribute('data-id', id);
-    btn.innerHTML = img;
-
-    return btn;
-  }
-
 
   return { createProject };
