+++ src/view/tasks.js
@@ -3,6 +3,9 @@
   This module manages changes to the DOM relevant to projects
 */
 
+import createTitle from './shared/title.js'
+import createButton from './shared/button.js'
+
 const tasks = (function () {
 
   function createTask(data) {
@@ -11,7 +14,22 @@ const tasks = (function () {
     task.classList.add('task');
     task.setAttribute('data-type', 'task');
     task.setAttribute('data-id', data.id);
-    task.innerText = data.title;
+
+    const title = createTitle(
+      data.title,
+      'task',
+      data.id
+      );
+
+    let flagIcon = (data.priority) ? 'flag' : 'outlined_flag'
+
+    const priorityBtn = createButton(
+      `<span class="material-icons" data-id="${data.id}">${flagIcon}</span>`,
+      'priority',
+      data.id
+      );
+
+    task.append(title, priorityBtn);
 
     return task
   }
